+++ app/src/main/java/woowacourse/shopping/presentation/products/ui/viewholder/ProductsItemViewHolder.kt
@@ -1,8 +1,9 @@
-package woowacourse.shopping.presentation.products.ui
+package woowacourse.shopping.presentation.products.ui.viewholder
 
 import androidx.annotation.CallSuper
 import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.presentation.products.ui.ProductsItem
 
 abstract class ProductsItemViewHolder<ITEM : ProductsItem, BINDING : ViewDataBinding>(
     protected val binding: BINDING,
