+++ app/src/main/res/layout/item_load_more.xml
@@ -10,7 +10,7 @@
 
         <variable
             name="onClickHandler"
-            type="woowacourse.shopping.presentation.products.ui.LoadMoreViewHolder.OnClickHandler" />
+            type="woowacourse.shopping.presentation.products.ui.viewholder.LoadMoreViewHolder.OnClickHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
